requirements.txt
rx
redis
nltk
python-dotenv
newspaper3k
Jinja2
GoogleNews